# contracts/openapi.yaml
components:
  schemas:
    AgentCreate:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        enabled:
          default: true
          title: Enabled
          type: boolean
        instructions:
          items:
            type: string
          title: Instructions
          type: array
        key:
          description: Unique stable key
          maxLength: 128
          minLength: 1
          pattern: ^[A-Za-z0-9][A-Za-z0-9._-]*$
          title: Key
          type: string
        knowledge_refs:
          items:
            type: string
          title: Knowledge Refs
          type: array
        model_ref:
          description: e.g. openai-compatible:gpt-5-mini
          maxLength: 256
          minLength: 1
          title: Model Ref
          type: string
        name:
          maxLength: 256
          minLength: 1
          title: Name
          type: string
        output_schema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output Schema
        role:
          maxLength: 256
          minLength: 1
          title: Role
          type: string
        skill_refs:
          items:
            type: string
          title: Skill Refs
          type: array
        tool_refs:
          items:
            type: string
          title: Tool Refs
          type: array
      required:
      - key
      - name
      - role
      - model_ref
      title: AgentCreate
      type: object
...
    AgentUpdate:
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        enabled:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
        expected_version:
          description: Optimistic lock version
          minimum: 1.0
          title: Expected Version
          type: integer
        instructions:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Instructions
        knowledge_refs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Knowledge Refs
        model_ref:
          anyOf:
          - maxLength: 256
            minLength: 1
            type: string
          - type: 'null'
          title: Model Ref
        name:
          anyOf:
          - maxLength: 256
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        output_schema:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Output Schema
        role:
          anyOf:
          - maxLength: 256
            minLength: 1
            type: string
          - type: 'null'
          title: Role
        skill_refs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Skill Refs
        tool_refs:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tool Refs
      required:
      - expected_version
      title: AgentUpdate
      type: object
...
    ErrorResponse:
      description: Stable error envelope returned by every ULOO Core endpoint.
      properties:
        code:
          description: Stable machine-readable error code
          title: Code
          type: string
        details:
          anyOf:
          - {}
          - type: 'null'
          title: Details
        message:
          description: Safe human-readable error message
          title: Message
          type: string
        request_id:
          title: Request Id
          type: string
        trace_id:
          title: Trace Id
          type: string
      required:
      - code
      - message
      - request_id
      - trace_id
      title: ErrorResponse
      type: object
...
    Page_AgentResponse_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/AgentResponse'
          title: Items
          type: array
        limit:
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          minimum: 0.0
          title: Offset
          type: integer
        total:
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - offset
      - limit
      title: Page[AgentResponse]
      type: object
...
info:
  title: ULOO Core
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/agents:
    get:
      description: List and search agents.
      operationId: list_agents_api_v1_agents_get
      parameters:
      - description: Search in name/key/role
        in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Search in name/key/role
          title: Q
      - in: query
        name: enabled
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Enabled
      - in: query
        name: model_ref
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Ref
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_AgentResponse_'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: List Agents
      tags:
      - agents
    post:
      description: Create a new Agent definition.
      operationId: create_agent_api_v1_agents_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Create Agent
      tags:
      - agents
  /api/v1/agents/{agent_id}:
    delete:
      description: Soft delete agent. Fails if agent is used by an enabled team.
      operationId: delete_agent_api_v1_agents__agent_id__delete
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Delete Agent
      tags:
      - agents
    get:
      description: Get agent details by ID.
      operationId: get_agent_api_v1_agents__agent_id__get
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Get Agent
      tags:
      - agents
    patch:
      description: Update agent with optimistic locking.
      operationId: update_agent_api_v1_agents__agent_id__patch
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdate'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Update Agent
      tags:
      - agents
  /api/v1/agents/{agent_id}/test-runs:
    post:
      description: Execute the persisted definition through the existing Agno adapter.
      operationId: test_run_agent_api_v1_agents__agent_id__test_runs_post
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/AgentTestRunRequest'
              - type: 'null'
              title: Body
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentTestRunResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Test Run Agent
      tags:
      - agents
  /api/v1/agents/{agent_id}/validate:
    post:
      description: Validate agent model and tool configuration without creating a Run.
      operationId: validate_agent_api_v1_agents__agent_id__validate_post
      parameters:
      - in: path
        name: agent_id
        required: true
        schema:
          format: uuid
          title: Agent Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - in: header
        name: x-uloo-workspace
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Uloo-Workspace
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentValidateResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Service Unavailable
        '504':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Gateway Timeout
      summary: Validate Agent
      tags:
      - agents
...